requirements.txt
requests~=2.31.0
kafka~=1.3.5
cassandra-driver~=3.29.1
pyspark~=3.5.1
confluent-kafka==2.3.0
apache-airflow==2.9.0
psycopg-binary
kafka-python
